HHH-16744 Wrong class created when reusing a join column for different entities in multiple ManyToOne
