+++ hibernate-core/src/main/java/org/hibernate/sql/results/graph/entity/internal/BatchEntityInsideEmbeddableSelectFetchInitializer.java
@@ -63,6 +63,7 @@ public class BatchEntityInsideEmbeddableSelectFetchInitializer extends AbstractB
 
 	@Override
 	public void resolveInstance(RowProcessingState rowProcessingState) {
+		resolveKey( rowProcessingState, referencedModelPart, parentAccess );
 		if ( entityKey == null ) {
 			return;
 		}
